ref: move session controller into user module
